Add thread-safe vector store
